refactor : filterWrapper 리팩터링
