+++ src/constant/filterButtonTitle.ts
@@ -1,0 +1,10 @@
+const POSITION_TITLES = {
+  cctv: 'CCTV',
+  safetyfacility: '안전시설',
+  firestation: '소방서',
+  heatshelter: '무더위 쉼터',
+  safetycenter: '안전센터',
+  emergencybell: '비상벨',
+};
+
+export default POSITION_TITLES;
