+++ src/utils/commonUtils.ts
@@ -1,0 +1,16 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const throttle = (func: (...args: any[]) => void, delay: number) => {
+  let throttled = false;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return (...args: any[]) => {
+    if (!throttled) {
+      throttled = true;
+      setTimeout(() => {
+        func(...args);
+        throttled = false;
+      }, delay);
+    }
+  };
+};
+
+export default throttle;
